implementing acceptTrade user story, still buggy

+++ TradeHub.Api/Models/DTOs/OfferDTOs.cs
@@ -3,9 +3,9 @@ namespace TradeHub.API.Models.DTOs
 {
     public class OfferDTO
     {
-        public int Id { get; set; }
-        public int UserId { get; set; }
-        public int TradeId { get; set; }
+        public long Id { get; set; }
+        public long UserId { get; set; }
+        public long TradeId { get; set; }
         public DateTimeOffset Created { get; set; }
     }
 }
