+++ TradeHub.Api/Repository/Interfaces/IOfferRepository.cs
@@ -4,8 +4,8 @@ namespace TradeHub.Api.Repository.Interfaces;
 
 public interface IOfferRepository
 {
-    Task<IEnumerable<Offer>> GetAllOffersInTradeAsync(int tradeId);
-    Task<Offer?> GetOfferAsync(int offerId);
+    Task<IEnumerable<Offer>> GetAllOffersInTradeAsync(long tradeId);
+    Task<Offer?> GetOfferAsync(long offerId);
     Task<Offer> CreateOfferAsync(Offer offer);
     Task<bool> UpdateOfferAsync(Offer offer);
     Task<bool> DeleteOfferAsync(Offer offer);
